Refactored comments a bit. To add comments to a page that represents an object of some sort, do something like this:
{% with species as object %}
    {% include "comments/_comments.html" %}
{% endwith %}

Replace species in the above with whatever variable represents the thing that the page is about (and that the comments should be attached to).

The above will give you a list of comments and a comment form (for logged in users) or a message suggesting that people log in (and which will redirect them back to the page they were on after they log in).